Fixes several issues related to git-reflow setup (#236)
There are times when a legacy user of git-reflow goes to re-run setup only to find themselves with blank configurations.  Since we are only validating presence of the github.oauth-token, there's an edge-case where the user could have _only_ the oauth-token configuration with missing user.  This causes some authentication failures and a bad user-experience.  This ensures that the user is asked for all required information, that the information is fully persisted to git-config, and gracefully handles authentication errors from GitHub in a more user-friendly way. Oh, and more tests :-)

Fixes #193