Fix tests and bump Ruby versions in test workflows
I'm not entirely certain when this broke, but at some point an issue was introduced where the Workflow gets loaded before the top-level `GitReflow` module functions (namely `logger`).  This fixes the tests and bumps the Ruby versions we want to test against in the GitHub workflow.

Merges #254
LGTM given by: @nhance